+++ Makefile
@@ -1,22 +1,36 @@
-CC=gcc
-FILES=src/core/*.c src/renderer/*.c
-DEBUG=src/debug/*.c
-TARGET=target/psx
-INCLUDES=./include
+CC	:=gcc
 
-WARNINGS=-Wall -Wextra 
-IGNORE_WARNINGS=-Wno-type-limits -Wno-unused-function -Wno-sign-compare -Wno-unused-parameter
-LIBRARIES=-lm -lSDL2 -lreadline
-FLAGS=$(WARNINGS) $(IGNORE_WARNINGS)
-DEBUGFLAGS=-g -pg -fsanitize=address 
+_DIR_INC     := include
+_DIR_SRC     := src
+_DIR_MODULES := core debug renderer
+_DIR_BUILD   := build
 
-run-debug:
-	clear 
-	make debug
-	./target/psx misc/SCPH1001.BIN .
+INCLUDE := -I$(_DIR_INC)
+SOURCES := $(foreach module,$(_DIR_MODULES),$(wildcard $(_DIR_SRC)/$(module)/*.c))
+OBJECTS := $(patsubst $(_DIR_SRC)/%.c,$(_DIR_BUILD)/%.o,$(SOURCES))
+TARGET  := $(_DIR_BUILD)/psx
+
+WARNINGS        := -Wall -Wextra 
+IGNORE_WARNINGS := -Wno-type-limits -Wno-unused-function -Wno-sign-compare -Wno-unused-parameter
+LIBRARIES       := -lm -lSDL2 -lreadline # -lubsan
+DEBUGFLAGS      := -g #-pg -fsanitize=undefined
+CFLAGS          := $(WARNINGS) $(IGNORE_WARNINGS) $(INCLUDE) $(DEBUGFLAGS)
+
+$(shell mkdir -p $(addprefix $(_DIR_BUILD)/, $(_DIR_MODULES)))
+
+$(TARGET): $(OBJECTS)
+	$(CC) $(OBJECTS) -o $@ $(LIBRARIES)
+
+$(_DIR_BUILD)/%.o: $(_DIR_SRC)/%.c
+	$(CC) $(CFLAGS) -o $@ -c $<
+
+.PHONY: clean run debug
+
+run:
+	./$(TARGET) misc/SCPH1001.BIN .
 
 debug:
-	$(CC) $(FILES) $(DEBUG) -o $(TARGET) -I$(INCLUDES) $(FLAGS) $(DEBUGFLAGS) $(LIBRARIES)
+	gdb --args ./$(TARGET) misc/SCPH1001.BIN .
 
-psx:
-	$(CC) $(FILES) -o $(TARGET) -I$(INCLUDES) $(FLAGS) $(LIBRARIES) -O3
+clean:
+	rm -rf $(_DIR_BUILD)
